refactor : common 패키지를 가독성있게 세부분리
