+++ src/main/java/cart/cart/exception/CartPersistenceFailedException.java
@@ -1,7 +1,7 @@
 package cart.cart.exception;
 
-import cart.common.PersistenceException;
-import cart.common.PersistenceExceptionMessages;
+import cart.common.exception.PersistenceException;
+import cart.common.exception.PersistenceExceptionMessages;
 
 public class CartPersistenceFailedException extends PersistenceException {
     public CartPersistenceFailedException(PersistenceExceptionMessages message) {
